Added Roster functionality
This commit adds a new Roster class to represent the data from the roster
of employees. The primary feature of the roster is that it includes data
associated with an individual employee's billable status and operational
unit.

Accordingly, the Timecard class is modified to merge the Roster data into
the timecards attribute.